requirements.txt
pytest
boto3
mock
python-packer
click
